weird error with glBindVertexArray after moving model loading code into SpatializeApp
