+++ spatialize/source/SpatializeApp.cpp
@@ -7,17 +7,39 @@
  */
 
 #include <SpatializeApp.h>
-#include <glm/glm.hpp>
-#include <glm/gtc/matrix_access.hpp>
-#include <glm/gtc/matrix_transform.hpp>
-#include <glm/ext.hpp>
 #include "example/ExampleCube.h"
-#include "VRModel.h"
 #include "MVRCore/StringUtils.H"
+#include <SOIL.h>
+#include "VRModel.h"
+#include <unistd.h>
 
 using namespace MinVR;
 
 namespace Spatialize {
+
+GLint TextureFromFile(const char* path, std::string directory)
+{
+     //Generate texture ID and load texture data 
+    std::string filename = string(path);
+    filename = directory + '/' + filename;
+    GLuint textureID;
+    glGenTextures(1, &textureID);
+    int width,height;
+    unsigned char* image = SOIL_load_image(filename.c_str(), &width, &height, 0, SOIL_LOAD_RGB);
+    // Assign texture to ID
+    glBindTexture(GL_TEXTURE_2D, textureID);
+    glTexImage2D(GL_TEXTURE_2D, 0, GL_RGB, width, height, 0, GL_RGB, GL_UNSIGNED_BYTE, image);
+    glGenerateMipmap(GL_TEXTURE_2D);    
+
+    // Parameters
+    glTexParameteri( GL_TEXTURE_2D, GL_TEXTURE_WRAP_S, GL_REPEAT );
+    glTexParameteri( GL_TEXTURE_2D, GL_TEXTURE_WRAP_T, GL_REPEAT );
+    glTexParameteri( GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_LINEAR_MIPMAP_LINEAR );
+    glTexParameteri( GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_LINEAR);
+    glBindTexture(GL_TEXTURE_2D, 0);
+    SOIL_free_image_data(image);
+    return textureID;
+}
 
 SpatializeApp::SpatializeApp(GLchar *path = NULL) : MinVR::AbstractMVRApp() {
 	_startTime = -1;
@@ -30,9 +52,137 @@ SpatializeApp::SpatializeApp(GLchar *path = NULL) : MinVR::AbstractMVRApp() {
 	_tempScale = 1.0f;
 	_scale = 1.0f;
     _path = path;
+    this->loadModel(_path);
 }
 
 SpatializeApp::~SpatializeApp() {
+}
+
+void SpatializeApp::loadModel(std::string path) {
+    Assimp::Importer importer;
+    const aiScene *scene = importer.ReadFile(path, aiProcess_Triangulate | aiProcess_FlipUVs);
+
+    if(!scene || scene->mFlags == AI_SCENE_FLAGS_INCOMPLETE || !scene->mRootNode) 
+    {
+        cout << "ERROR::ASSIMP::" << importer.GetErrorString() << endl;
+        return;
+    }
+    this->directory = path.substr(0, path.find_last_of('/'));
+
+    this->processNode(scene->mRootNode, scene);
+
+    sleep(10);
+}
+
+void SpatializeApp::processNode(aiNode *node, const aiScene *scene) {
+    for (GLuint i = 0; i < node->mNumMeshes; i++) {
+        aiMesh *mesh = scene->mMeshes[node->mMeshes[i]];
+        this->_meshes.push_back(this->processMesh(mesh, scene));
+    }
+
+    for (GLuint i = 0; i < node->mNumChildren; i++) {
+        this->processNode(node->mChildren[i], scene);
+    }
+}
+
+Mesh SpatializeApp::processMesh(aiMesh *mesh, const aiScene *scene) {
+    vector<Vertex> vertices;
+    vector<GLuint> indices;
+    vector<Texture> textures;
+
+    for (GLuint i = 0; i < mesh->mNumVertices; i++) {
+        Vertex vertex;
+
+        glm::vec3 vect;
+        vect.x = mesh->mVertices[i].x;
+        vect.y = mesh->mVertices[i].y;
+        vect.z = mesh->mVertices[i].z;
+        vertex.Position = vect;
+
+        if (!min.x && !min.y && !min.z && !max.x && !max.y && !max.z)
+        {
+            min = vect;
+            max = vect;
+        }
+        else
+        {
+            if (vect.x < min.x) { min.x = vect.x; }
+            if (vect.y < min.y) { min.y = vect.y; }
+            if (vect.z < min.z) { min.z = vect.z; }
+            if (vect.x > max.x) { max.x = vect.x; }
+            if (vect.y > max.y) { max.y = vect.y; }
+            if (vect.z > max.z) { max.z = vect.z; }
+        }
+
+        vect.x = mesh->mNormals[i].x;
+        vect.y = mesh->mNormals[i].y;
+        vect.z = mesh->mNormals[i].z;
+
+        vertex.Normal = vect;
+
+        if(mesh->mTextureCoords[0]) // Does the mesh contain texture coordinates?
+        {
+            glm::vec2 vec;
+            vec.x = mesh->mTextureCoords[0][i].x; 
+            vec.y = mesh->mTextureCoords[0][i].y;
+            vertex.TexCoords = vec;
+        }
+        else
+            vertex.TexCoords = glm::vec2(0.0f, 0.0f);  
+
+        vertices.push_back(vertex);
+    }
+
+    for(GLuint i = 0; i < mesh->mNumFaces; i++)
+    {
+        aiFace face = mesh->mFaces[i];
+        // Retrieve all indices of the face and store them in the indices vector
+        for(GLuint j = 0; j < face.mNumIndices; j++)
+            indices.push_back(face.mIndices[j]);
+    }
+
+    if (mesh->mMaterialIndex >= 0) {
+        aiMaterial *material = scene->mMaterials[mesh->mMaterialIndex];
+
+        vector<Texture> diffuseMaps = this->loadMaterialTextures(material, aiTextureType_DIFFUSE, "texture_diffuse");
+        textures.insert(textures.end(), diffuseMaps.begin(), diffuseMaps.end());
+
+        vector<Texture> specularMaps = this->loadMaterialTextures(material, aiTextureType_SPECULAR, "texture_specular");
+        textures.insert(textures.end(), specularMaps.begin(), specularMaps.end());
+    }
+
+    return Mesh(vertices, indices, textures);    
+}
+
+vector<Texture> SpatializeApp::loadMaterialTextures(aiMaterial* mat, aiTextureType type, std::string typeName)
+{
+    vector<Texture> textures;
+    for(GLuint i = 0; i < mat->GetTextureCount(type); i++)
+    {
+        aiString str;
+        mat->GetTexture(type, i, &str);
+        // Check if texture was loaded before and if so, continue to next iteration: skip loading a new texture
+        GLboolean skip = false;
+        for(GLuint j = 0; j < _textures_loaded.size(); j++)
+        {
+            if(_textures_loaded[j].path == str)
+            {
+                textures.push_back(_textures_loaded[j]);
+                skip = true; // A texture with the same filepath has already been loaded, continue to next one. (optimization)
+                break;
+            }
+        }
+        if(!skip)
+        {   // If texture hasn't been loaded already, load it
+            Texture texture;
+            texture.id = TextureFromFile(str.C_Str(), this->directory);
+            texture.type = typeName;
+            texture.path = str;
+            textures.push_back(texture);
+            this->_textures_loaded.push_back(texture);  // Store it as texture loaded for entire model, to ensure we won't unnecesery load duplicate textures.
+        }
+    }
+    return textures;
 }
 
 void SpatializeApp::doUserInputAndPreDrawComputation(
@@ -125,7 +275,7 @@ void SpatializeApp::initVBO(int threadId)
     if (!_path)
 	    _scene[threadId] = SceneRef(new ExampleCube());
     else
-        _scene[threadId] = SceneRef(new VRModel(_path));
+        _scene[threadId] = SceneRef(new VRModel(_meshes, min, max));
     _mutex.unlock();
 }
